Move remaining functions in libFuzzer's launcher.py to libfuzzer.py
Keep launcher.py around, to prevent breakages. This will be removed in a
future PR. Also add some logic to handle unset APP_PATHs in
get_command_line_for_application and get_crash_stacktrace_output to
prepare for that.
