+++ src/python/tests/core/bot/fuzzers/libFuzzer/libfuzzer_test_data/corrupted_stats.txt
