+++ src/python/tests/core/bot/fuzzers/libFuzzer/libfuzzer_test_data/startup_crash.txt
